requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-storage
google-cloud-batch
protobuf
authlib